Add files via upload
The chosen method used to solve the stabilisation issues of the videos was key point based stabilisation. This involves estimating the geometric transformations between pairs of consecutive frames from their corresponding points. Oriented FAST and Rotated BRIEF (ORB) image features extraction is used; in which FAST is used to find key-points, then Harris corner measure is applied to find the top points among them, also using a pyramid to produce multiscale features (. P. A., 2015).  To improve rotation invariance for computing orientation, moments were used for x and y. The BREIF feature contains a crucial property in which each bit feature has a large variance and mean near 0.5; as soon as it is oriented along key-point direction it loses this property and becomes more distributed. A Desired feature is to have the tests uncorrelated to which ORB achieves this by running a greedy search amongst all possible binary tests to find the ones that have both high variance and are uncorrelated .